+++ src/py_types/mod.rs
@@ -1,9 +1,12 @@
+mod converter;
 mod errors;
 mod param;
 mod rows;
 mod types;
+mod value;
 
 // re-export
 pub use errors::{py_error, DBError, PySQLxError, PySQLxInvalidParamError};
 pub use rows::{PySQLxColumnTypes, PySQLxResponse, PySQLxRow, PySQLxRows};
-pub use types::{PySQLxStatement, PySQLxValue};
+pub use types::PySQLxStatement;
+pub use value::PySQLxValue;
