+++ src/py_types/value.rs
@@ -1,0 +1,65 @@
+use bigdecimal::BigDecimal;
+use chrono::{DateTime, NaiveDate, NaiveTime, Utc};
+use serde::{Deserialize, Deserializer, Serialize, Serializer};
+use serde_json::Value as JsonValue;
+use uuid::Uuid;
+
+// this type is a placeholder for the actual type
+type PyValueArray = Vec<PySQLxValue>;
+
+#[derive(Debug, Clone, PartialEq, Deserialize, Serialize)]
+#[serde(untagged)]
+pub enum PySQLxValue {
+    // true, false
+    Boolean(bool),
+    // text
+    String(String),
+    // red, green, blue
+    Enum(String),
+    // [red, green, blue]
+    EnumArray(Vec<String>),
+    // 1.0
+    Int(i64),
+    // Vec<String>,
+    Array(PyValueArray),
+    // { "name": "foo", "age": 42 }
+    Json(JsonValue),
+    // <body>...</body>
+    Xml(String),
+    // 00000000-0000-0000-0000-000000000000
+    Uuid(Uuid),
+    // 00:00:00
+    Time(NaiveTime),
+    // 2020-01-01
+    Date(NaiveDate),
+    // 2020-01-01T00:00:01
+    DateTime(DateTime<Utc>),
+    // 18373737.8274
+    Float(f64),
+    // Vec<u8>
+    Bytes(Vec<u8>),
+    // REPRESENT A BIG DECIMAL
+    #[serde(
+        serialize_with = "serialize_bigdecimal",
+        deserialize_with = "deserialize_bigdecimal"
+    )]
+    Numeric(BigDecimal),
+    // None
+    Null,
+}
+
+fn serialize_bigdecimal<S>(value: &BigDecimal, serializer: S) -> Result<S::Ok, S::Error>
+where
+    S: Serializer,
+{
+    serializer.serialize_str(&value.to_string())
+}
+
+fn deserialize_bigdecimal<'de, D>(deserializer: D) -> Result<BigDecimal, D::Error>
+where
+    D: Deserializer<'de>,
+{
+    let s = String::deserialize(deserializer)?;
+    BigDecimal::parse_bytes(s.as_bytes(), 10)
+        .ok_or_else(|| serde::de::Error::custom("Invalid BigDecimal"))
+}
